requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
dateparser==1.2.0
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.8.4
Flask==3.0.3
